Split GitLab item dialog
